fix(abi): Input_Default is uint8_t, not c_int (frontier_screen)
Companion to the significantDigits fix. C declares `uint8_t Input_Default`
(c47.h:501) and every other owner binds it u8, but frontier_screen had it as
c_int. It is read in the same displayable-value guard --
`Input_Default == ID_DP` -- so as a 4-byte c_int the three high bytes come from
adjacent globals and the equality can spuriously fail even when the real 1-byte
value is ID_DP.

Verified: zig build sim and zig build test exit 0 (9674 testSuite), fmt clean.

+++ zig_src/frontier/frontier_screen.zig
@@ -1381,7 +1381,7 @@ inline fn COPY_REGISTER_STRING_TO(dest: [*c]u8, regist: calcRegister_t) void {
 // ---------------------------------------------------------------------------
 extern var significantDigits: u8;
 extern var exponentLimit: i16;
-extern var Input_Default: c_int;
+extern var Input_Default: u8;
 inline fn checkHP() bool {
     return significantDigits <= 16 and displayStack == 1 and exponentLimit == 99 and Input_Default == ID_DP and (calcMode == CM_NORMAL or calcMode == CM_NIM);
 }
